stepOne_1

+++ src/main/java/ru/skypro/homework/controller/AdsController.java
@@ -1,5 +1,6 @@
 package ru.skypro.homework.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -25,31 +26,63 @@ public class AdsController {
 
     @PostMapping
     public ResponseEntity<AdDto> addAd(@RequestParam Object properties, @RequestParam MultipartFile image) {
-        return ResponseEntity.ok(adsService.addAd(properties, image));
+        if () {
+            return ResponseEntity.ok(adsService.addAd(properties, image));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<ExtendedAdDto> getInfo(@PathVariable Integer id) {
-        return ResponseEntity.ok(adsService.getInfoExtendedAd(id));
+        if () {
+            return ResponseEntity.ok(adsService.getInfoExtendedAd(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+//404 Not found????????
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<AdDto> deleteAd(@PathVariable Integer id) {
-        return ResponseEntity.ok(adsService.deleteAd(id));
+        if () {
+            return ResponseEntity.ok(adsService.deleteAd(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+//Forbidden????
+//Not found????
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<AdDto> patchAd(@PathVariable Integer id) {
-        return ResponseEntity.ok(adsService.patchAd(id));
+        if () {
+            return ResponseEntity.ok(adsService.patchAd(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+//Forbidden????
+//Not found????
     }
 
     @GetMapping("/me")
     public ResponseEntity<AdsDto> getUserAds() {
-        return ResponseEntity.ok(adsService.getUserAds());
+        if () {
+            return ResponseEntity.ok(adsService.getUserAds());
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
     }
 
     @PatchMapping("/{id}/image")
     public ResponseEntity<String> patchImage(@PathVariable Integer id) {
-        return ResponseEntity.ok(adsService.patchImage(id));
+        if () {
+            return ResponseEntity.ok(adsService.patchImage(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
     }
 }
